enrollments linked to user should be in users routes
